fixed: last (right) swipe now displays final msg
