+++ src/main_page/keep_or_delete.html
@@ -152,6 +152,8 @@
 
 </body>
 <script type="module" src="../settings.js"></script>
+<script type="module" src="../rename.js"></script>
+<script type="module" src="../sort.js"></script>
 <script type="module" src="../currentIndex.js"></script>
 <script type="module" src="./keep_or_delete.js"></script>
